refactor: Import errors from errors/index
