+++ src/errors/validation.error.ts
@@ -1,9 +1,9 @@
-import ApplicationError from './application.error';
-import HttpStatusCodes from 'http-status-codes';
+import { ApplicationError } from './application.error';
+import { StatusCodes } from 'http-status-codes';
 
 // todo: customize based on validation package
 export default class ValidationError extends ApplicationError {
 	constructor(message: string = 'Validation failed') {
-		super(message, HttpStatusCodes.BAD_REQUEST);
+		super(message, StatusCodes.BAD_REQUEST);
 	}
 }
