+++ src/middlewares/notFound.middleware.ts
@@ -1,9 +1,8 @@
 import { NextFunction, Request, Response } from 'express';
-import ApplicationError from '../errors/application.error';
-import HttpStatusCodes from 'http-status-codes';
-import ErrMessages from '../errors/error-messages';
+import { ApplicationError, ErrMessages } from '../errors';
+import { StatusCodes } from 'http-status-codes';
 
 export function NotFoundHandler(req: Request, res: Response, _next: NextFunction) {
 	const message = `${req.method} ${req.path}`;
-	throw new ApplicationError(ErrMessages.http.NotFound, HttpStatusCodes.NOT_FOUND);
+	throw new ApplicationError(ErrMessages.http.NotFound, StatusCodes.NOT_FOUND);
 }
